Update all-in-one json CI test
The new implementation supports tags/endpoints/passthru. This
change updates the CI test to exercise a full all-in-one json,
with endpoints, tags and passthru args included.

Coverage is only for oslat-k8s-alma8 scenario and should be
expanded for other benchmarks/endpoints/userenvs.

This is not a replacament for mv-parms param mode tests as it has
been added as a complimentary test for all-in-one json (from-file)
mode.